improve music pages loading
